+++ Exception/SignedUriException.php
@@ -11,9 +11,13 @@
 
 namespace Symfony\Component\HttpFoundation\Exception;
 
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Attribute\WithHttpStatus;
+
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
+#[WithHttpStatus(Response::HTTP_NOT_FOUND)]
 abstract class SignedUriException extends \RuntimeException implements ExceptionInterface
 {
 }